Good object and currency helper methods

+++ src/main/java/it/cajani/matteo/salestaxes/GoodCategory.java
@@ -3,17 +3,19 @@ package it.cajani.matteo.salestaxes;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @AllArgsConstructor
 public enum GoodCategory {
 
-    BOOKS(0, 5),
-    FOOD(0, 5),
-    HEALTH(0, 5),
-    AUDIO(10, 5),
-    BEAUTY(10, 5);
+    BOOKS(BigDecimal.valueOf(0), BigDecimal.valueOf(5)),
+    FOOD(BigDecimal.valueOf(0), BigDecimal.valueOf(5)),
+    HEALTH(BigDecimal.valueOf(0), BigDecimal.valueOf(5)),
+    AUDIO(BigDecimal.valueOf(10), BigDecimal.valueOf(5)),
+    BEAUTY(BigDecimal.valueOf(10), BigDecimal.valueOf(5));
 
-    private final int baseTax;
-    private final int importTax;
+    private final BigDecimal baseTax;
+    private final BigDecimal importTax;
 
 }
